+++ src/Data/DbSetting.cs
@@ -1,0 +1,13 @@
+﻿namespace uConfig.Data;
+public record DbSetting
+{
+	public string Key { get; set; }
+	public string Value { get; set; }
+
+	public DbSetting() { }
+	public DbSetting(string key, string value)
+	{
+		this.Key = key;
+		this.Value = value;
+	}
+}